requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.2
Werkzeug==3.0.2
Flask-SQLAlchemy==3.1.1
mysql-connector-python==8.4.0
pymongo==4.7.1
python-dotenv==1.0.1
Flask-SeaSurf==1.1.1